Accept large values for UnsignedData.age
While running the example project against my account, I got a room event
that's over 3 years old. The program failed to deserialize it because
the value is too large for an `int` (2^31 milliseconds is less than
a month).

This commit changes the field to `long`, which can represent millions of
years.